df-init/function_requirements.txt
google-api-python-client==1.7.4
google-cloud-logging==1.6.0
grpcio==1.10.0
google-auth
